Crate solana_gossip

Modules

cluster_info

The cluster_info module defines a data structure that is shared by all the nodes in the network over a gossip control plane. The goal is to share small bits of off-chain information and detect and repair partitions.

cluster_info_metrics

contact_info

crds

This module implements Cluster Replicated Data Store for asynchronous updates in a distributed network.

crds_entry

crds_gossip

Crds Gossip.

crds_gossip_error

crds_gossip_pull

Crds Gossip Pull overlay.

crds_gossip_push

Crds Gossip Push overlay.

crds_shards

crds_value

deprecated 🔒

duplicate_shred

duplicate_shred_handler

duplicate_shred_listener

epoch_slots

gossip_error

gossip_service

The gossip_service module implements the network control plane.

legacy_contact_info

ping_pong

push_active_set 🔒

received_cache 🔒

weighted_shuffle

The weighted_shuffle module provides an iterator over shuffled weights.

Macros

socketaddr

socketaddr_any