Module solana_gossip::cluster_info

The cluster_info module defines a data structure that is shared by all the nodes in the network over a gossip control plane. The goal is to share small bits of off-chain information and detect and repair partitions.

This CRDT only supports a very limited set of types. A map of Pubkey -> Versioned Struct. The last version is always picked during an update.

The network is arranged in layers:

layer 0 - Leader.

layer 1 - As many nodes as we can fit

layer 2 - Everyone else, if layer 1 is 2^10, layer 2 should be able to fit 2^20 number of nodes.

Bank needs to provide an interface for us to query the stake weight

Macros

set_socket 🔒

set_socket 🔒

Structs

ClusterInfo

Node

PruneData 🔒

PullData 🔒

SignData 🔒

Sockets

__FieldVisitor 🔒

__FieldVisitor 🔒

__Visitor 🔒

__Visitor 🔒

__Visitor 🔒

__Visitor 🔒

__Visitor 🔒

__Visitor 🔒

Enums

ClusterInfoError

Protocol 🔒

__Field 🔒

__Field 🔒

Constants

BYTES_PER_INTERVAL 🔒

CRDS_UNIQUE_PUBKEY_CAPACITY 🔒

DATA_PLANE_FANOUT

The Data plane fanout size, also used as the neighborhood size

DEFAULT_CONTACT_DEBUG_INTERVAL_MILLIS

DEFAULT_CONTACT_SAVE_INTERVAL_MILLIS

DEFAULT_EPOCH_DURATION_MS 🔒

DUPLICATE_SHRED_MAX_PAYLOAD_SIZE 🔒

FIELDS 🔒

GOSSIP_PING_CACHE_CAPACITY 🔒

GOSSIP_PING_CACHE_RATE_LIMIT_DELAY 🔒

GOSSIP_PING_CACHE_TTL 🔒

GOSSIP_PING_TOKEN_SIZE 🔒

Number of bytes in the randomly generated token sent with ping messages.

GOSSIP_SLEEP_MILLIS

milliseconds we sleep for between gossip requests

INTERVAL_MS 🔒

MAX_ACCOUNTS_HASHES

Maximum number of hashes in AccountsHashes a node publishes such that the serialized size of the push/pull message stays below PACKET_DATA_SIZE.

MAX_BLOOM_SIZE

The maximum size of a bloom filter

MAX_BUDGET_MULTIPLE 🔒

MAX_CRDS_OBJECT_SIZE

MAX_GOSSIP_TRAFFIC 🔒

A hard limit on incoming gossip messages Chosen to be able to handle 1Gbps of pure gossip traffic 128MB/PACKET_DATA_SIZE

MAX_INCREMENTAL_SNAPSHOT_HASHES

Maximum number of incremental hashes in SnapshotHashes a node publishes such that the serialized size of the push/pull message stays below PACKET_DATA_SIZE.

MAX_LEGACY_SNAPSHOT_HASHES

Maximum number of hashes in LegacySnapshotHashes a node publishes such that the serialized size of the push/pull message stays below PACKET_DATA_SIZE.

MAX_PRUNE_DATA_NODES 🔒

Maximum number of origin nodes that a PruneData may contain, such that the serialized size of the PruneMessage stays below PACKET_DATA_SIZE.

MINIMUM_VALIDATOR_PORT_RANGE_WIDTH

MIN_NUM_STAKED_NODES 🔒

Minimum number of staked nodes for enforcing stakes in gossip.

MIN_STAKE_FOR_GOSSIP 🔒

Minimum stake that a node should have so that its CRDS values are propagated through gossip (few types are exempted).

PULL_RESPONSE_MIN_SERIALIZED_SIZE 🔒

Minimum serialized size of a Protocol::PullResponse packet.

PUSH_MESSAGE_MAX_PAYLOAD_SIZE 🔒

Max size of serialized crds-values in a Protocol::PushMessage packet. This is equal to PACKET_DATA_SIZE minus serialized size of an empty push message: Protocol::PushMessage(Pubkey::default(), Vec::default())

RECV_TIMEOUT 🔒

RECV_TIMEOUT 🔒

SUBMIT_GOSSIP_STATS_INTERVAL 🔒

THROTTLE_DELAY 🔒

VALIDATOR_PORT_RANGE

VARIANTS 🔒

Functions

compute_retransmit_peers

Turbine logic 1 - For the current node find out if it is in layer 1 1.1 - If yes, then broadcast to all layer 1 nodes 1 - using the layer 1 index, broadcast to all layer 2 nodes assuming you know neighborhood size 1.2 - If no, then figure out what layer the node is in and who the neighbors are and only broadcast to them 1 - also check if there are nodes in the next layer and repeat the layer 1 to layer 2 logic Returns Neighbor Nodes and Children Nodes (neighbors, children) for a given node based on its stake

count_packets_received 🔒

extend 🔒

filter_on_shred_version 🔒

get_epoch_duration 🔒

make_accounts_hashes_message

merge 🔒

push_messages_to_peer

retain_staked 🔒

Type Definitions

Ping 🔒