Module solana_gossip::crds_value

Structs

AccountsHashes

CrdsValue

CrdsValue that is replicated across the cluster

LegacyVersion

LowestSlot

NodeInstance

SnapshotHashes

Version

Vote

Vote ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

Enums

CrdsData

CrdsData that defines the different types of items CrdsValues can hold

CrdsValueLabel

Type of the replicated value These are labels for values in a record that is associated with Pubkey

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

Constants

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

FIELDS ğŸ”’

MAX_EPOCH_SLOTS

MAX_SLOT

MAX_VOTES

MAX_WALLCLOCK

VARIANTS ğŸ”’

Functions

filter_current ğŸ”’

Filters out an iterator of crds values, returning the unique ones with the most recent wallclock.

new_rand_timestamp ğŸ”’

Random timestamp for tests and benchmarks.

sanitize_wallclock ğŸ”’

Type Definitions

EpochSlotsIndex

LegacySnapshotHashes

VoteIndex